requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
absl-py==1.3.0
    # via
    #   tensorboard
    #   tensorflow
accelerate==0.13.0
    # via -r requirements.in
aiohttp==3.8.3
    # via
    #   datasets
    #   fsspec
aiosignal==1.3.1
    # via aiohttp
astunparse==1.6.3
    # via tensorflow
async-timeout==4.0.2
    # via aiohttp
attrs==22.1.0
    # via
    #   aiohttp
    #   jsonlines
blis==0.7.9
    # via thinc
cachetools==5.2.0
    # via google-auth
catalogue==2.0.8
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via
    #   aiohttp
    #   requests
ckiptagger==0.2.1
    # via -r requirements.in
click==8.1.3
    # via
    #   nltk
    #   typer
confection==0.0.3
    # via thinc
cymem==2.0.7
    # via
    #   preshed
    #   spacy
    #   thinc
datasets==2.5.2
    # via -r requirements.in
dill==0.3.5.1
    # via
    #   datasets
    #   multiprocess
filelock==3.8.0
    # via
    #   huggingface-hub
    #   transformers
flatbuffers==22.10.26
    # via tensorflow
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2022.10.0
    # via datasets
gast==0.4.0
    # via tensorflow
google-auth==2.14.1
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
google-pasta==0.2.0
    # via tensorflow
grpcio==1.50.0
    # via
    #   tensorboard
    #   tensorflow
h5py==3.7.0
    # via tensorflow
huggingface-hub==0.10.1
    # via
    #   datasets
    #   transformers
idna==3.4
    # via
    #   requests
    #   yarl
importlib-metadata==5.0.0
    # via markdown
jinja2==3.1.2
    # via spacy
joblib==1.2.0
    # via nltk
jsonlines==3.1.0
    # via -r requirements.in
keras==2.10.0
    # via tensorflow
keras-preprocessing==1.1.2
    # via tensorflow
langcodes==3.3.0
    # via spacy
libclang==14.0.6
    # via tensorflow
markdown==3.4.1
    # via tensorboard
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.13
    # via
    #   -r requirements.in
    #   datasets
murmurhash==1.0.9
    # via
    #   preshed
    #   spacy
    #   thinc
nltk==3.7
    # via -r requirements.in
numpy==1.23.4
    # via
    #   accelerate
    #   blis
    #   datasets
    #   h5py
    #   keras-preprocessing
    #   opt-einsum
    #   pandas
    #   pyarrow
    #   spacy
    #   tensorboard
    #   tensorflow
    #   thinc
    #   transformers
oauthlib==3.2.2
    # via requests-oauthlib
opt-einsum==3.3.0
    # via tensorflow
packaging==21.3
    # via
    #   accelerate
    #   datasets
    #   huggingface-hub
    #   spacy
    #   tensorflow
    #   transformers
pandas==1.5.1
    # via
    #   -r requirements.in
    #   datasets
pathy==0.6.2
    # via spacy
preshed==3.0.8
    # via
    #   spacy
    #   thinc
protobuf==3.19.6
    # via
    #   tensorboard
    #   tensorflow
psutil==5.9.4
    # via accelerate
pyarrow==10.0.0
    # via datasets
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pydantic==1.10.2
    # via
    #   confection
    #   spacy
    #   thinc
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via pandas
pytz==2022.6
    # via pandas
pyyaml==6.0
    # via
    #   accelerate
    #   huggingface-hub
    #   transformers
regex==2022.10.31
    # via
    #   nltk
    #   transformers
requests==2.28.1
    # via
    #   datasets
    #   fsspec
    #   huggingface-hub
    #   requests-oauthlib
    #   responses
    #   spacy
    #   tensorboard
    #   transformers
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
responses==0.18.0
    # via datasets
rouge==1.0.1
    # via -r requirements.in
rsa==4.9
    # via google-auth
sentencepiece==0.1.97
    # via -r requirements.in
six==1.16.0
    # via
    #   astunparse
    #   google-auth
    #   google-pasta
    #   grpcio
    #   keras-preprocessing
    #   python-dateutil
    #   rouge
    #   tensorflow
smart-open==5.2.1
    # via pathy
spacy==3.4.2
    # via -r requirements.in
spacy-legacy==3.0.10
    # via spacy
spacy-loggers==1.0.3
    # via spacy
srsly==2.4.5
    # via
    #   confection
    #   spacy
    #   thinc
tensorboard==2.10.1
    # via tensorflow
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.1
    # via tensorboard
tensorflow==2.10.0
    # via -r requirements.in
tensorflow-estimator==2.10.0
    # via tensorflow
tensorflow-io-gcs-filesystem==0.27.0
    # via tensorflow
termcolor==2.1.0
    # via tensorflow
thinc==8.1.5
    # via spacy
tokenizers==0.13.2
    # via transformers
torch==1.12.1
    # via accelerate
tqdm==4.64.1
    # via
    #   -r requirements.in
    #   datasets
    #   huggingface-hub
    #   nltk
    #   spacy
    #   transformers
transformers==4.24.0
    # via -r requirements.in
typer==0.4.2
    # via
    #   pathy
    #   spacy
typing-extensions==4.4.0
    # via
    #   huggingface-hub
    #   pydantic
    #   tensorflow
    #   torch
urllib3==1.26.12
    # via
    #   requests
    #   responses
wasabi==0.10.1
    # via
    #   spacy
    #   spacy-loggers
    #   thinc
werkzeug==2.2.2
    # via tensorboard
wheel==0.38.3
    # via
    #   astunparse
    #   tensorboard
wrapt==1.14.1
    # via tensorflow
xxhash==3.1.0
    # via datasets
yarl==1.8.1
    # via aiohttp
zipp==3.10.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
